requirements.txt
streamlit>=1.20
paddleocr>=2.7
torch>=2.0
transformers>=4.30
opencv-python
pillow
pdf2image
PyMuPDF
tqdm
editdistance
ollama>=0.5.0
requests
reportlab
